Address PR review feedback: move JsonConverter guard and relocate tests
Move the JsonConverterAttribute check from IsArrayElementCompatible into IsTokenCompatibleWithType, where it protects all callers rather than just the array peeking path. Add a regression test for SumType element arrays (e.g., SumType<string, int>[]) that would fail without this guard.

Move SumConverterTypeFilteringTests from the Razor test project to ProtocolUnitTests/Converters, colocating them with the product code they test. Fix license header and file encoding to match repo conventions.
