+++ src/LanguageServer/Protocol/Protocol/Converters/SumConverter.cs
@@ -287,6 +287,14 @@ internal sealed class SumConverter<T> : JsonConverter<T>
 
     private static bool IsTokenCompatibleWithType(ref Utf8JsonReader reader, Type type)
     {
+        // If the type has a custom JsonConverter (e.g., SumType elements), we can't
+        // reason about token compatibility since the converter may accept multiple
+        // token types. Treat as compatible and let the normal try/catch handle it.
+        if (type.GetCustomAttribute<JsonConverterAttribute>() is not null)
+        {
+            return true;
+        }
+
         return reader.TokenType switch
         {
             JsonTokenType.True or JsonTokenType.False
@@ -341,14 +349,6 @@ internal sealed class SumConverter<T> : JsonConverter<T>
         }
 
         var elementType = arrayType.GetElementType()!;
-
-        // If the element type has a custom JsonConverter (e.g., SumType elements),
-        // we can't reason about token compatibility since the converter may accept
-        // multiple token types. Let the normal try/catch handle it.
-        if (elementType.GetCustomAttribute<JsonConverterAttribute>() is not null)
-        {
-            return true;
-        }
 
         // Peek at first array element to disambiguate array types.
         var peekReader = reader;
