+++ src/LanguageServer/ProtocolUnitTests/Converters/SumConverterTypeFilteringTests.cs
@@ -1,14 +1,16 @@
-// Licensed to the .NET Foundation under one or more agreements.
+﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
 
 using System;
 using System.Collections.Generic;
 using System.Runtime.ExceptionServices;
 using System.Text.Json;
 using System.Text.Json.Serialization;
+using Roslyn.LanguageServer.Protocol;
 using Xunit;
 
-namespace Microsoft.CodeAnalysis.Razor.Serialization;
+namespace Microsoft.CodeAnalysis.LanguageServer.UnitTests.Converters;
 
 /// <summary>
 /// Tests for SumConverter's type filtering logic, which avoids costly exception-based
